Extract Ergodox default visualizer into simple_visualizer
It's good enough if you only want to change the lcd text and color.
